// .PRD/erp-refactor/v3/types/supplier-quotation.ts
import { z } from 'zod';

export const SupplierQuotationSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  title: z.string().optional().default('{supplier_name}'),
  naming_series: z.enum(['PUR-SQTN-.YYYY.-']),
  supplier: z.string(),
  supplier_name: z.string().optional(),
  company: z.string(),
  status: z.enum(['Draft', 'Submitted', 'Stopped', 'Cancelled', 'Expired']),
  transaction_date: z.string().default('Today'),
  valid_till: z.string().optional(),
  quotation_number: z.string().optional(),
  has_unit_price_items: z.boolean().optional().default(false),
  amended_from: z.string().optional(),
  cost_center: z.string().optional(),
  project: z.string().optional(),
  currency: z.string(),
  conversion_rate: z.number(),
  buying_price_list: z.string().optional(),
  price_list_currency: z.string().optional(),
  plc_conversion_rate: z.number().optional(),
  ignore_pricing_rule: z.boolean().optional().default(false),
  items: z.array(z.unknown()),
  total_qty: z.number().optional(),
  total_net_weight: z.number().optional(),
  base_total: z.number().optional(),
  base_net_total: z.number().optional(),
  total: z.number().optional(),
  net_total: z.number().optional(),
  tax_category: z.string().optional(),
  taxes_and_charges: z.string().optional(),
  shipping_rule: z.string().optional(),
  incoterm: z.string().optional(),
  named_place: z.string().optional(),
  taxes: z.array(z.unknown()).optional(),
  base_taxes_and_charges_added: z.number().optional(),
  base_taxes_and_charges_deducted: z.number().optional(),
  base_total_taxes_and_charges: z.number().optional(),
  taxes_and_charges_added: z.number().optional(),
  taxes_and_charges_deducted: z.number().optional(),
  total_taxes_and_charges: z.number().optional(),
  apply_discount_on: z.enum(['Grand Total', 'Net Total']).optional().default('Grand Total'),
  base_discount_amount: z.number().optional(),
  additional_discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  base_grand_total: z.number().optional(),
  base_rounding_adjustment: z.number().optional(),
  base_rounded_total: z.number().optional(),
  base_in_words: z.string().max(240).optional(),
  grand_total: z.number().optional(),
  rounding_adjustment: z.number().optional(),
  rounded_total: z.number().optional(),
  in_words: z.string().max(240).optional(),
  disable_rounded_total: z.boolean().optional().default(false),
  other_charges_calculation: z.string().optional(),
  item_wise_tax_details: z.array(z.unknown()).optional(),
  pricing_rules: z.array(z.unknown()).optional(),
  supplier_address: z.string().optional(),
  address_display: z.string().optional(),
  contact_person: z.string().optional(),
  contact_display: z.string().optional(),
  contact_mobile: z.string().optional(),
  contact_email: z.string().email().optional(),
  shipping_address: z.string().optional(),
  shipping_address_display: z.string().optional(),
  billing_address: z.string().optional(),
  billing_address_display: z.string().optional(),
  tc_name: z.string().optional(),
  terms: z.string().optional(),
  letter_head: z.string().optional(),
  group_same_items: z.boolean().optional().default(false),
  select_print_heading: z.string().optional(),
  language: z.string().optional(),
  auto_repeat: z.string().optional(),
  is_subcontracted: z.boolean().optional().default(false),
  opportunity: z.string().optional(),
});

export type SupplierQuotation = z.infer<typeof SupplierQuotationSchema>;

export const SupplierQuotationInsertSchema = SupplierQuotationSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type SupplierQuotationInsert = z.infer<typeof SupplierQuotationInsertSchema>;